Chore: Refactor software tests. Naming things.
