+++ tests/zyp/moksha/test_jq.py
@@ -5,9 +5,6 @@ https://github.com/jqlang/jq/blob/master/src/builtin.jq
 
 from copy import deepcopy
 
-import pytest
-from jmespath.exceptions import ParseError
-
 from zyp.model.moksha import MokshaRule, MokshaTransformation
 
 
@@ -22,7 +19,7 @@ def test_moksha_jq_idempotency():
 
 def test_moksha_jq_select_pick_keys():
     """
-    Verify selecting elements with moksha/jq.
+    Verify selecting elements with moksha/jq, by picking individual keys from objects.
     """
     data_in = [{"meta": {"id": "Hotzenplotz", "timestamp": 123456789}, "data": {"abc": 123, "def": 456}}]
     data_out = [{"meta": {"id": "Hotzenplotz"}, "data": {"abc": 123}}]
@@ -32,7 +29,7 @@ def test_moksha_jq_select_pick_keys():
 
 def test_moksha_jq_select_pick_indices():
     """
-    Verify selecting elements with moksha/jq.
+    Verify selecting elements with moksha/jq, by picking individual indices from arrays.
     """
     data_in = [{"data": [1, {"foo": "bar"}, 2]}]
     data_out = [{"data": [1, 2]}]
@@ -40,9 +37,9 @@ def test_moksha_jq_select_pick_indices():
     assert transformation.apply(data_in) == data_out
 
 
-def test_moksha_jq_select_drop_keys():
+def test_moksha_jq_select_drop_keys_object_direct():
     """
-    Verify selecting elements with moksha/jq.
+    Verify selecting elements with moksha/jq, by dropping individual keys from objects.
     """
     data_in = [{"meta": {"id": "Hotzenplotz", "timestamp": 123456789}, "data": {"abc": 123, "def": 456}}]
     data_out = [{"meta": {"id": "Hotzenplotz"}, "data": {"abc": 123}}]
@@ -52,7 +49,7 @@ def test_moksha_jq_select_drop_keys():
 
 def test_moksha_jq_select_drop_indices():
     """
-    Verify selecting elements with moksha/jq.
+    Verify selecting elements with moksha/jq, by dropping individual indices from arrays.
     """
     data_in = [{"data": [1, {"foo": "bar"}, 2]}]
     data_out = [{"data": [1, 2]}]
@@ -113,7 +110,7 @@ def test_moksha_jq_cast_array_iterable():
     assert transformation.apply(data_in) == data_out
 
 
-def test_moksha_jq_cast_dict():
+def test_moksha_jq_cast_object():
     """
     Verify type casting using moksha/jq.
     """
@@ -147,63 +144,3 @@ def test_moksha_jq_flatten_array():
 
 def test_moksha_jq_rule_disabled():
     assert MokshaRule(type="jq", expression=". | tostring", disabled=True).compile().evaluate(42.42) == 42.42
-
-
-def test_transon_duplicate_records():
-    """
-    Verify record duplication works well.
-    """
-    data_in = [{"foo": "bar", "baz": "qux"}]
-    data_out = [{"foo": "bar", "baz": "qux"}] * 42
-    transformation = MokshaTransformation().transon({"$": "expr", "op": "mul", "value": 42})
-    assert transformation.apply(data_in) == data_out
-
-
-def test_transon_idempotency():
-    """
-    Idempotent transformations should not modify data.
-    """
-    data = [{"foo": "bar"}, {"baz": "qux"}]
-    transformation = MokshaTransformation().transon({"$": "this"})
-    assert transformation.apply(deepcopy(data)) == data
-
-
-def test_moksha_rule():
-    moksha = MokshaRule(type="jmes", expression="@").compile()
-    assert moksha.transformer.expression == "@"
-    assert moksha.transformer.parsed == {"type": "current", "children": []}
-
-
-def test_moksha_runtime_rule_success():
-    assert MokshaRule(type="jmes", expression="@").compile().evaluate(42.42) == 42.42
-
-
-def test_moksha_runtime_rule_syntax_error():
-    with pytest.raises(ParseError) as ex:
-        MokshaRule(type="jmes", expression="@foo").compile()
-    assert ex.match("Unexpected token: foo")
-
-
-def test_moksha_runtime_rule_invalid_transformer():
-    rule = MokshaRule(type="jmes", expression="@").compile()
-    rule.transformer = "foo"
-    with pytest.raises(TypeError) as ex:
-        rule.evaluate(42.42)
-    assert ex.match("Evaluation failed. Type must be either jmes or jq or transon: foo")
-
-
-def test_moksha_empty():
-    """
-    Empty JSON Pointer expression means "root node".
-    """
-    with pytest.raises(ValueError) as ex:
-        MokshaTransformation().jmes("")
-    assert ex.match("JMESPath expression cannot be empty")
-
-    with pytest.raises(ValueError) as ex:
-        MokshaTransformation().jq("")
-    assert ex.match("jq expression cannot be empty")
-
-    with pytest.raises(ValueError) as ex:
-        MokshaTransformation().transon("")
-    assert ex.match("transon expression cannot be empty")
